rag-daemon/requirements.txt
# ============================================================
# Hybrid RAG Edge Daemon — Dependencies
# ============================================================
# Install with: pip install -r requirements.txt
#
# Python 3.11+ is required.
# ============================================================

# --- Web Framework -------------------------------------------------------
fastapi>=0.111.0,<0.200.0          # Async HTTP server framework
uvicorn[standard]>=0.29.0,<0.30.0  # ASGI server (includes websockets, watchfiles)

# --- Data Validation & Settings ------------------------------------------
pydantic>=2.7.0,<3.0.0             # Data models and validation
pydantic-settings>=2.2.0,<3.0.0    # .env / config.json loading via BaseSettings

# --- Vector Database -----------------------------------------------------
chromadb>=0.5.0,<0.6.0             # Local persistent vector store

# --- Embeddings ----------------------------------------------------------
sentence-transformers>=3.0.0,<4.0.0  # Local embedding models (all-MiniLM-L6-v2)

# --- Utilities -----------------------------------------------------------
python-dotenv>=1.0.0,<2.0.0        # .env file support (used by pydantic-settings)
requests>=2.31.0,<3.0.0            # HTTP client for Spotify Web API
